requirements.txt
asgiref==3.3.4
cached-property==1.5.2
dj-database-url==0.5.0
Django==3.1.6
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
enum-compat==0.0.3
gunicorn==20.1.0
psycopg2==2.8.6
python-dotenv==0.17.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
whitenoise==5.2.0
